+++ Layout 2.1/LoginPage.aspx
@@ -1,67 +1,11 @@
-﻿<%--<%@ Page Language="C#" AutoEventWireup="true" CodeBehind="Login.aspx.cs" Inherits="Login_Page.WebForm1" %>
-
-<!DOCTYPE html>
-
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head runat="server">
-    <title> </title>
-    <style>
-        h1
-        {
-            text-align:center;
-        }
-
-        .div
-        {
-            margin:100px 600px;
-            padding:20px;
-            border: 1px double;
-        }
-        .txt
-        
-        {
-            width:300px;
-            height:30px;
-
-        }
-        .btn
-        {
-            height:40px;
-            width:300px;
-            margin-top:10px;
-
-        }
-    </style>
-</head>
-<body>
-    <form id="form1" runat="server">
-        <h1> Login Page</h1>
-        <hr />
-        <div class="div">
-            USERNAME:<br />
-            <asp:TextBox CssClass="txt" ID="textuser" runat="server"/> <br />
-
-            PASSWORD:<br />
-             <asp:TextBox CssClass="txt" ID="textpass" runat="server"/> <br />
-
-            <asp:Button CssClass="btn" ID="sbt" Text="Submit" runat="server" OnClick="sbt_Click" />
-
-        </div>
-    </form>
-</body>
-</html>--%>
-
-
-
-<%@ Page Language="C#" AutoEventWireup="true" CodeBehind="WebForm1.aspx.cs" Inherits="LoginAnimated.WebForm1" %>
+﻿<%@ Page Language="C#" AutoEventWireup="true" CodeBehind="LoginPage.aspx.cs" Inherits="Layout_2._1.WebForm1" %>
 
 <!DOCTYPE html>
 
 <html xmlns="http://www.w3.org/1999/xhtml">
 <head runat="server">
     <title></title>
-    
-<link href="animatedLoginCSS2.css" rel="stylesheet" />
+<link href="animatedLoginCSS1.css" rel="stylesheet" />
     <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@500&display=swap" rel="stylesheet">
     
 </head>
@@ -91,10 +35,12 @@
             <asp:Label ID="pass" runat="server" Text="" ForeColor="Red" Font-Size="10px"></asp:Label>
         </div>
         <asp:Button ID="Submit" CssClass="Submit" runat="server" Text="Login" onclick="Submit_Click" BackColor="#CCCCCC" />
-        
+        <br />
+        <asp:Label ID="loginfail" CssClass="loginfail" runat="server" Text="" ForeColor="#FF3300"></asp:Label>
 
             
     </form>
     </div>
 </body>
 </html>
+
