+++ Layout 2.1/Set Offs 2 new backend.csproj
@@ -67,11 +67,15 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Content Include="5259.jpg" />
+    <Content Include="animatedLoginCSS1.css" />
     <Content Include="cal.jpg" />
     <Content Include="Calendar 1.aspx" />
     <Content Include="Create_Abs.aspx" />
-    <Content Include="Login.aspx" />
+    <Content Include="flexur2.jpg" />
+    <Content Include="undraw_pic_profile_re_7g2h.svg" />
     <Content Include="Web.config" />
+    <Content Include="LoginPage.aspx" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Calendar 1.aspx.cs">
@@ -89,14 +93,14 @@
       <DependentUpon>Create_Abs.aspx</DependentUpon>
     </Compile>
     <Compile Include="DBConnection.cs" />
-    <Compile Include="Login.aspx.cs">
-      <DependentUpon>Login.aspx</DependentUpon>
+    <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="LoginPage.aspx.cs">
+      <DependentUpon>LoginPage.aspx</DependentUpon>
       <SubType>ASPXCodeBehind</SubType>
     </Compile>
-    <Compile Include="Login.aspx.designer.cs">
-      <DependentUpon>Login.aspx</DependentUpon>
+    <Compile Include="LoginPage.aspx.designer.cs">
+      <DependentUpon>LoginPage.aspx</DependentUpon>
     </Compile>
-    <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
