+++ Layout 2.1/animatedLoginCSS1.css
@@ -1,0 +1,157 @@
+
+
+
+* {
+    margin:0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: 'Poppins', sans-serif;
+}
+
+body {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-height: 100vh; 
+    border:solid 3px;
+   
+   
+}
+
+
+.box {
+    position: absolute;
+    width: 400px;
+    height: 480px;
+    background-color: #f8fbff;
+    border-radius: 8px;
+    right: 60px;
+    overflow: hidden;
+}
+    .box form {
+        position: absolute;
+        inset: 4px;
+        background: #f8fbff;
+        padding: 20px 25px 20px 30px;
+        border-radius: 8px;
+        z-index: 2;
+        display: flex;
+        flex-direction: column;
+    }
+.box form h2
+{
+    color:black;
+    font-weight:500;
+    text-align:center;
+    letter-spacing:0.1em;
+
+
+}
+.box form .inputBox
+{
+    position:relative;
+    width:300px;
+   
+    
+}
+    .box form .Img
+    {
+        margin: 0px 0px 0px 100px;
+
+    }
+/*.box form .inputBox input
+{
+    position:relative;
+    width:100%;
+    padding: 20px 10px 10px;
+    background:transparent;
+    outline:none;
+    box-shadow: none;
+    border:none;
+    color:#23242a;
+    font-size:1em;
+    letter-spacing:0.05em;
+    transition:0.5s;
+    z-index:10;
+
+
+}
+
+    .box form.inputBox input:valid ~ .textbox,
+    .box form.inputBox input:focus ~ .textbox
+
+    {
+        color:#fff;
+        font-size:0.75em;
+        transform:translateY(-34px);
+    }
+    /* .box form .A
+    {
+       height:40px;
+       width:80px;
+       background-color:white;
+
+    }  */
+    .Submit {
+        border: none;
+        outline: none;
+        padding: 9px 25px;
+        background: gray;
+        cursor: pointer;
+        font-size: 0.9em;
+        border-radius: 4px;
+        font-weight: 600;
+        width: 100px;
+        margin-top: 40px;
+        margin-left: 100px;
+    }
+
+.box::before {
+    content: '';
+    position: absolute;
+    top: -50%;
+    left: -50%;
+    width: 380px;
+    height: 420px;
+    background: linear-gradient(0deg,transparent,transparent,Red,Red,Red);
+    z-index: 1;
+    transform-origin: bottom right;
+    animation: animate 6s linear infinite;
+}
+@keyframes animate
+{
+    0%
+    {
+        transform:rotate(0deg);
+    }
+    100% {
+        transform: rotate(360deg);
+    }
+}
+.backimg{
+    position:fixed;
+    z-index:-1;
+    width:80%;
+    height:100%;
+    left:-100px;
+    bottom:-80px;
+   
+   
+
+   
+}
+.backimg1 {
+    position: fixed;
+   
+    height:60px;
+    width:180px;
+    top:50px;
+    left:50px;
+   
+    background-color:transparent;
+   
+}
+.box form .loginfail 
+{
+    right: 10px;
+}
